Apply row traversal policy to component preset table
